% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmdb-package.R
\docType{package}
\name{hmdb-package}
\alias{hmdb}
\alias{hmdb-package}
\title{hmdb: Summary Data from the Human Metabolome Database}
\description{
Data from the Human Metabolome Database.
}
\author{
\strong{Maintainer}: William Oldham \email{wm.oldham@gmail.com} (\href{https://orcid.org/0000-0003-3029-4866}{ORCID})

}
\keyword{internal}
