% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmdb_proteins.R
\docType{data}
\name{hmdb_proteins}
\alias{hmdb_proteins}
\title{HMDB metabolite-protein interactions}
\format{
A data frame with 839,671 rows and 3 columns:
\describe{
\item{hmdb}{HMDB accession number}
\item{name}{metabolite name}
\item{protein}{protein gene symbol}
}
}
\source{
\url{https://hmdb.ca/system/downloads/current/hmdb_metabolites.zip}
}
\usage{
hmdb_proteins
}
\description{
A list of metabolite-protein interactions extracted from the Human Metabolome
Database. Only proteins with HUGO Gene Nomenclature Committee symbols were
included in the output. Preparation details may be found in the \code{data-raw}
directory of the package source.
}
\keyword{datasets}
